Add general patches
