+++ CatQuest_Randomizer/Patches/QuestPatches.cs
@@ -5,19 +5,6 @@ using System.Collections.Generic;
 
 namespace CatQuest_Randomizer.Patches
 {
-    [BepInPlugin("enable.logger", "EnableLogger", "1.0.0")]
-    public class EnableLogger : BaseUnityPlugin
-    {
-        private static ManualLogSource Logger;
-
-        private void Awake()
-        {
-            Logger = base.Logger;
-            Harmony harmony = new Harmony("enable.logger");
-            harmony.PatchAll();
-        }
-    }
-
     [HarmonyPatch(typeof(Quest), nameof(Quest.End))]
     public class QuestEndPatch
     {
